[#767] Factored out DropdownToggle
- Simplifies LanguageToggle by using DropdownToggle for shared logic
- Enhances maintainability with a dedicated DropdownToggle component
- Removes redundant code and improves event handling
- Updates DefaultFooter to include JurisdictionToggle component
